+++ hlambda-vscode/api/static-fallback/index.js
@@ -1,0 +1,5 @@
+module.exports = async (req, res) => {
+	res.status(404);
+	res.setHeader('Cache-Control', 'no-store');
+	res.send('Not Found');
+};